% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asHtml.R
\name{as.html}
\alias{as.html}
\title{Transform to html}
\usage{
as.html(h)
}
\arguments{
\item{h}{string}
}
\value{
HTML form of the string for easy insertion in Shiny or Rmarkdown documents
}
\description{
Transform string into HTML ready for Rmarkdown and Shiny
}
\examples{
string <- "<b>smile</b>" #html b tag included to make the string bold
string_trans <- as.html(string)
cat(string) #when inserted in Rmarkdown will not show up as bolded
cat(string_trans) #when inserted into Rmarkdown will show up as bolded after knitting

}
