% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isEmpty.R
\name{is.empty}
\alias{is.empty}
\title{Is string empty}
\usage{
is.empty(string)
}
\arguments{
\item{string}{the string to be checked}
}
\value{
Returns TRUE or FALSE
}
\description{
Return TRUE or FALSE after checking if a string is empty
}
\examples{
string <- "r2resume"
string2 <- NULL
is.empty(string)
is.empty(string2)
}
