% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{sym}
\alias{sym}
\title{Get a desired symbol}
\usage{
sym(
  ...,
  font.size = NULL,
  font.weight = NULL,
  font.color = NULL,
  if.error = FALSE,
  search.units.each = FALSE
)
}
\arguments{
\item{...}{the symbols to display}

\item{font.size}{font size as integer}

\item{font.weight}{font weight as integer}

\item{font.color}{font color e.g. red, blue, cyan}

\item{if.error}{error response}

\item{search.units.each}{whether to search single letter columns}
}
\value{
symbol hex for HTML displays
}
\description{
Get a symbol by name or number for use in Rmarkdown or Shiny apps
}
\examples{
sym("trademark")
sym("alpha")
sym("arrow-right")
sym(144)

}
