% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{symbol}
\alias{symbol}
\title{Get a desired symbol}
\usage{
symbol(
  ...,
  font.size = NULL,
  font.weight = NULL,
  font.color = NULL,
  if.error = FALSE,
  search.units.each = FALSE
)
}
\arguments{
\item{...}{the symbols to display}

\item{font.size}{font size as integer}

\item{font.weight}{font weight as integer}

\item{font.color}{font color e.g. red, blue, cyan}

\item{if.error}{error response}

\item{search.units.each}{whether to search single letter columns}
}
\value{
symbol hex for HTML integration
}
\description{
Get a symbol by name or numberfor use in Rmarkdown or Shiny apps, could be a mathematical symbol or other kinds of symbols
}
\section{Complete cheatcheat for symbols}{

Common symbols include alpha, beta, copyright, registered, theta, sigma. See the complete list at
\url{https://github.com/oobianom/r2symbols}.
}

\examples{
symbol("trademark")
symbol("alpha")
symbol("arrow-right")
symbol(144)

}
