% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolSetting.R
\name{symbol.setting}
\alias{symbol.setting}
\title{Symbol setting}
\usage{
symbol.setting(font.size = NULL, font.weight = NULL, font.color = NULL)
}
\arguments{
\item{font.size}{font size as integer}

\item{font.weight}{font weight as integer}

\item{font.color}{font color e.g. red, blue, cyan}
}
\value{
No values returned
}
\description{
Font styling for symbols
}
\examples{
symbol.setting(40, "bold", "green") # ascribe settings for all symbols
symbol("alpha") # display the apha symbol
symbol.setting() # clear settings for symbols
symbol("alpha") # dislay the alpha symbol
}
