% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolUnits.R
\name{symbol.units}
\alias{symbol.units}
\title{Transform custom units}
\usage{
symbol.units(units, error = FALSE)
}
\arguments{
\item{units}{the desired unit e.g ug/mL,pg/mL, ab, yL}

\item{error}{whether to report any errors}
}
\value{
Transformed units with substitution of symbols with appropriate symbol hex
}
\description{
Enter a unit and it will automatically be transformed to the appropriate form
}
\examples{
symbol.units("ug/mL") # u will be transformed to mu
symbol.units("pg/mL") # no letter will be transformed
}
