% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCat.R
\name{symCategory}
\alias{symCategory}
\title{Fetch symbols by category}
\usage{
symCategory(
  category,
  font.size = NULL,
  font.weight = NULL,
  font.color = NULL,
  if.error = FALSE,
  search.units.each = FALSE
)
}
\arguments{
\item{category}{symbol categories to fetch}

\item{font.size}{font size as integer}

\item{font.weight}{font weight as integer}

\item{font.color}{font color e.g. red, blue, cyan}

\item{if.error}{error response}

\item{search.units.each}{whether to search single letter columns}
}
\value{
get symbols from a particular category
}
\description{
Fetch one or more symbols by category
}
\section{Complete cheatcheat for symbols}{

Common symbols include alpha, beta, copyright, registered, theta, sigma. See the complete list at
\url{https://github.com/oobianom/r2symbols}.
}

\examples{
symCategory("arrows")
symCategory("smiley")
symCategory("music")
symCategory("random")

}
