% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchByKeyCat.R
\name{symKey}
\alias{symKey}
\title{Fetch multiple symbols by keyword}
\usage{
symKey(
  keyword,
  font.size = NULL,
  font.weight = NULL,
  font.color = NULL,
  if.error = TRUE
)
}
\arguments{
\item{keyword}{symbol keyword to fetch}

\item{font.size}{font size as integer}

\item{font.weight}{font weight as integer}

\item{font.color}{font color e.g. red, blue, cyan}

\item{if.error}{show error response}
}
\value{
get group of symbols maptching a particular keyword
}
\description{
Fetch one or more symbols by keyword
}
\section{Complete cheatcheat for symbols}{

Common symbols include alpha, beta, copyright, registered, theta, sigma. See the complete list at
\url{https://github.com/oobianom/r2symbols}.
}

\examples{
symKey("arrow", font.size = 45)
symKey("animal", font.color = "red")
symKey("random454")
symKey("show nothing")

}
