% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symload.R
\name{symload}
\alias{symload}
\title{Symbol repository loader}
\usage{
symload(
  font.size = NULL,
  font.weight = NULL,
  font.color = NULL,
  if.error = TRUE
)
}
\arguments{
\item{font.size}{font size as integer}

\item{font.weight}{font weight as integer}

\item{font.color}{font color e.g. red, blue, cyan}

\item{if.error}{error response}
}
\value{
No returns, but symbols are loaded
}
\description{
Loads all the available symbols
}
\examples{
symload(40, "normal", "purple") # load symbols and add symbol settings
symload() # load symbols without custom settings

}
