% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeScripts.R
\name{trans.r2symbols}
\alias{trans.r2symbols}
\title{Transform HTML into symbols}
\usage{
trans.r2symbols(
  font.size = NULL,
  font.weight = c("normal", "bold", "lighter"),
  font.color = NULL
)
}
\arguments{
\item{font.size}{font size as integer}

\item{font.weight}{font weight eg. normal, bold or lighter}

\item{font.color}{font color e.g. red, blue, cyan}
}
\value{
A rendered HTML where simple text is transformed to symbols using HTML codes
}
\description{
Transform texts in HTML to symbols
}
\note{
This function simple inserts a script within the output HTML, which transforms the content using the relevant HTML codes.
}
\examples{
trans.r2symbols()
}
