% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeScripts.R
\name{use.r2symbols}
\alias{use.r2symbols}
\title{Transform HTML into symbols}
\usage{
use.r2symbols()
}
\value{
A rendered HTML where simple text is transformed to symbols using HTML codes
}
\description{
Transform texts in HTML to symbols
}
\note{
This function simple inserts a script within the output HTML, which transforms the content using the relevant HTML codes.
}
\examples{
use.r2symbols()

}
